galaxys2: remove the bootmode=2 cmdline parameter.
Even though the addition of the bootmode=2 cmdline parameter is supposed
to be conditional, this cmdline parameter is still always added, causing
issues with kernels that use the bootmode cmdline parameter added by the
bootloader in their init scripts to tell if a recovery ramdisk should be
used as the ramdisk.

For anyone having found this commit: if you're already able to boot your
kernel just fine, this change won't really have any effect for you.